dana-prototype/pyproject.toml
[tool.poetry]
name = "dana-funding"
version = "0.1.0"
description = "Dana AI Copilot - Deep AI agent for academic outreach and funding applications"
authors = ["CanApply <dev@canapply.ca>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
starlette = "^0.35.0"

# Database
prisma = "^0.12.0"

# HTTP Client
httpx = "^0.26.0"
aiohttp = "^3.9.0"

# AI/ML
openai = "^1.12.0"
tiktoken = "^0.6.0"

# AWS
aiobotocore = "^2.11.0"
boto3 = "^1.34.0"

# Data Processing
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"

# Async utilities
aiofiles = "^23.2.1"

# Document processing
python-docx = "^1.1.0"
pypdf2 = "^3.0.0"
markdown = "^3.5.0"

# Caching and hashing
blake3 = "^0.4.0"
redis = "^5.0.0"

# Utilities
python-multipart = "^0.0.9"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^24.1.0"
isort = "^5.13.0"
mypy = "^1.8.0"
ruff = "^0.2.0"

[tool.poetry.scripts]
dana = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | prisma
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = ["prisma", ".venv"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = ["prisma", "tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["prisma", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"





